chore: Update Module Dependencies (fixes #101) - docs: Update modularization-graph image
