+++ core/domain/build.gradle.kts
@@ -8,8 +8,6 @@ android {
 }
 
 dependencies {
-    implementation(projects.core.common)
-    implementation(projects.core.data)
-    implementation(projects.core.model)
-    implementation(projects.core.network)
+    api(projects.core.data)
+    api(projects.core.model)
 }
